fix(peach): rename parser file to match convention of other app directories
